==> R/Analysis RCT/2022/analysis_rct/01_preprocess_rct.Rmd <==
---
title: "Preprocess data for RCT analysis"
output:
  pdf_document:
    keep_tex:  true
    latex_engine: xelatex
  html_document: default
header-includes:
   - \usepackage{floatrow}
   - \floatsetup[figure]{capposition=top}
---

```{r echo=FALSE, message=FALSE, results='hide'}
rm(list=ls())
gc()
```

```{r echo=FALSE, message=FALSE, results='hide'}
library(rlang)
library(tidyr)
library(fastDummies)
library(reshape2)
library(corrplot)
library(stargazer)
library(stringr)
library(ggplot2)
library(rddtools)
library(parallel)
library(evd)
library(magic)
library(mvtnorm)
library(LaplacesDemon)
library(truncnorm)
library(compiler)
library(feather)
library(magrittr)
library(geosphere)
library(Hmisc)
library(plyr)
library(tidyr)
library(fastDummies)
library(reshape2)
library(corrplot)
library(stargazer)
library(stringr)
library(ggplot2)
#library(tidyverse)
library(haven)
library(MASS)
library(stringi)
library(openxlsx)
library(readxl)
library(kableExtra)
library(dplyr)
```

```{r echo=FALSE, message=FALSE, results='hide'}
my_colors <- RColorBrewer::brewer.pal(4, "Pastel1")[1:2]
myblue = '#2F5597'
```

```{r}
dropbox_dir <- "~/Dropbox/"
```

# Preprocess for Analysis RCT --------------------

## Read data and merge
```{r}
car <- read.csv(paste(dropbox_dir, "/Mistakes Structural/Code/Python/Cartillas/data/2022/carreras_v2.csv", sep=''), header=TRUE, sep=";")
dm_int <- read.csv(file = paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/outputs/2022/preprocessing_rct_interim.csv", sep=""), header=TRUE, sep=";")
dm_fin <- read.csv(file = paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/outputs/2022/preprocessing_rct_finales.csv", sep=""), header=TRUE, sep=";")
scores <- read.csv(file = paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/data/2022/puntajes_consolidado_v3.csv", sep=""), header=TRUE, sep=",")
colnames(dm_int)[2:length(colnames(dm_int))] <- paste(colnames(dm_int)[2:length(colnames(dm_int))], '_int', sep='')
colnames(dm_fin)[2:length(colnames(dm_fin))] <- paste(colnames(dm_fin)[2:length(colnames(dm_int))], '_fin', sep='')
```

```{r}
dm <- merge(dm_int, dm_fin, by="mrun")
```

```{r}
dm %<>% mutate(apps_before = (codigo_carrera_1_int > 0) + (codigo_carrera_2_int > 0) +
                             (codigo_carrera_3_int > 0) + (codigo_carrera_4_int > 0) +
                             (codigo_carrera_5_int > 0) + (codigo_carrera_6_int > 0) +
                             (codigo_carrera_7_int > 0) + (codigo_carrera_8_int > 0) +
                             (codigo_carrera_9_int > 0) + (codigo_carrera_10_int > 0),
              valid_apps_before = (marca_1_int == 25) + (marca_2_int == 25) +
                                  (marca_3_int == 25) + (marca_4_int == 25) +
                                  (marca_5_int == 25) + (marca_6_int == 25) +
                                  (marca_7_int == 25) + (marca_8_int == 25) +
                                  (marca_9_int == 25) + (marca_10_int == 25) ,
               apps_after = (codigo_carrera_1_fin > 0) + (codigo_carrera_2_fin > 0) +
                             (codigo_carrera_3_fin > 0) + (codigo_carrera_4_fin > 0) +
                             (codigo_carrera_5_fin > 0) + (codigo_carrera_6_fin > 0) +
                             (codigo_carrera_7_fin > 0) + (codigo_carrera_8_fin > 0) +
                             (codigo_carrera_9_fin > 0) + (codigo_carrera_10_fin > 0),
              valid_apps_after = (marca_1_fin == 25) + (marca_2_fin == 25) +
                                  (marca_3_fin == 25) + (marca_4_fin == 25) +
                                  (marca_5_fin == 25) + (marca_6_fin == 25) +
                                  (marca_7_fin == 25) + (marca_8_fin == 25) +
                                  (marca_9_fin == 25) + (marca_10_fin == 25) ,
              changed_apps = (codigo_carrera_1_int != codigo_carrera_1_fin) + (codigo_carrera_2_int != codigo_carrera_2_fin) +
                             (codigo_carrera_3_int != codigo_carrera_3_fin) + (codigo_carrera_4_int != codigo_carrera_4_fin) +
                             (codigo_carrera_5_int != codigo_carrera_5_fin) + (codigo_carrera_6_int != codigo_carrera_6_fin) +
                             (codigo_carrera_7_int != codigo_carrera_7_fin) + (codigo_carrera_8_int != codigo_carrera_8_fin) +
                             (codigo_carrera_9_int != codigo_carrera_9_fin) + (codigo_carrera_10_int != codigo_carrera_10_fin)
            )

```

```{r}
of <- read.csv(paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/data/2022/seleccion_20_01_22.csv", sep=''))
open <- read.csv(paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/data/2022/accesos-cartilla-postulaciones.csv", sep=''))
pumi <- read.csv(paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/data/2022/accesos-cartilla-puntajes.csv", sep=''))
pumi %<>% dplyr::rename(MRUN = id, pumi = ingreso)
ms <- read.csv(paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/outputs/2022/master_students_with_treatment_and_stratas.csv", sep=''), header=TRUE, sep=',')
#TODO: we need to check how the probability variables that are in the object ms were computed and if we have to just drop them now
ms %<>% select(!(starts_with("prob") | starts_with("overall")))

op <- open %>% group_by(distinct_id) %>% dplyr::summarise(t_open = min(time)) %>% mutate(open = 1) %>% dplyr::rename(MRUN = distinct_id)
df <- merge(ms, of, by.x="MRUN", by.y="mrun", all.x=TRUE)
df <- merge(df, dm, by.x="MRUN", by.y="mrun", all.x=TRUE)

df <- merge(df, op, by="MRUN", all.x=TRUE)
df$open[is.na(df$open)] <- 0

df <- merge(df, pumi, by="MRUN", all.x=TRUE)
df$pumi[is.na(df$pumi)] <- 0
```

```{r}
df %<>% mutate(assigned_oficial = ifelse((estado_pref_01 %in% c(24, 26)) + (estado_pref_02 %in% c(24, 26)) +
                          (estado_pref_03 %in% c(24, 26)) + (estado_pref_04 %in% c(24, 26)) +
                          (estado_pref_05 %in% c(24, 26)) + (estado_pref_06 %in% c(24, 26)) +
                          (estado_pref_07 %in% c(24, 26)) + (estado_pref_08 %in% c(24, 26)) +
                          (estado_pref_09 %in% c(24, 26)) + (estado_pref_10 %in% c(24, 26)) > 0, 1, 0))

df %<>% mutate(changed = ifelse(changed_apps > 0, 1, 0),
               increased = ifelse(apps_after > apps_before, 1, 0),
               decreased = ifelse(apps_after < apps_before, 1, 0))

df %<>% mutate(reduced_mistakes = ifelse(changed == 1 & valid_apps_before < valid_apps_after, 1, 0))

df %<>% mutate(increased_valid = ifelse(valid_apps_after > valid_apps_before, 1, 0),
               decreased_valid = ifelse(valid_apps_after < valid_apps_before, 1, 0))
```

## Preprocessing programs
```{r}
of <- read.csv(paste(dropbox_dir, "Mistakes Structural/Code/Python/Cartillas/data/2022/seleccion_20_01_22.csv", sep=''))
car <- read.csv(paste(dropbox_dir, "/Mistakes Structural/Code/Python/Cartillas/data/2022/carreras_v2.csv", sep=''), header=TRUE, sep=";")

of_l <- reshape(of[,c(1,4:53)], direction='long',
                    varying=colnames(of[,4:53]),
                    times=c('01', '02', '03', '04', '05', '06', '07', '08', '09', '10'),
                    v.names=c('cod_carrera_pref', 'estado_pref','ptje_pref', 'lugar_pref', 'pond_acad_pref'),
                    idvar='mrun')
colnames(of_l) <- c("mrun", "pref", "codigo_carrera", "marca", "orden", "puntano", "puntpond") # this is just me not knowing how to use reshape
# I can tell! JAJAJAJAJA

cutoffs <- of_l %>% filter(marca == 24) %>% group_by(codigo_carrera) %>% dplyr::summarise(num_admitted = n(), min_score = pmax(min(puntpond), 210.0))
#TODO: this is not taking into account that many programs do not fill their vacancies. We
# can assign max(min_pond, 210) for those that do not fill VACANTES_REG
car %<>% mutate(VACANTES_REG = VACANTES_1_SEMESTRE + VACANTES_2_SEMESTRE + SOBRECUPO_1_SEMESTRE + SOBRECUPO_2_SEMESTRE)
car <- merge(car, cutoffs, by.x="CODIGO_DEMRE", by.y="codigo_carrera", all.x=TRUE)
```

## Consolidate interim and final assignment
```{r}
reg_int <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/asignacion_obtenida_reg_interim_completando_con_finales.csv", sep=''), header=TRUE, sep=';')
bea_int <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/asignacion_obtenida_bea_interim_completando_con_finales.csv", sep=''), header=TRUE, sep=';')
pace_int <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/asignacion_obtenida_pace_interim_completando_con_finales.csv", sep=''), header=TRUE, sep=';')

reg_fin <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/asignacion_obtenida_reg_finales.csv", sep=''), header=TRUE, sep=';')
bea_fin <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/asignacion_obtenida_bea_finales.csv", sep=''), header=TRUE, sep=';')
pace_fin <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/asignacion_obtenida_pace_finales.csv", sep=''), header=TRUE, sep=';')
```

```{r}
pace_int_a <- pace_int %>% filter(marca == 24)
bea_int_a  <- bea_int %>% filter(marca == 24 & !(id_alumno %in% pace_int_a$id_alumno))
ass_int <- rbind(pace_int_a, bea_int_a)
ass_int <- rbind(ass_int, reg_int %>% filter(marca == 24 & !(id_alumno %in% ass_int$id_alumno) ))

pace_fin_a <- pace_fin %>% filter(marca == 24)
bea_fin_a  <- bea_fin %>% filter(marca == 24 & !(id_alumno %in% pace_fin_a$id_alumno))
ass_fin <- rbind(pace_fin_a, bea_fin_a)
ass_fin <- rbind(ass_fin, reg_fin %>% filter(marca == 24 & !(id_alumno %in% ass_fin$id_alumno) ))
```

```{r}
ass_int$assigned <- 1
ass_fin$assigned <- 1

rm(pace_int_a, bea_int_a, pace_fin_a, bea_fin_a)
```

```{r}
ass <- merge(ass_int %>% dplyr::select(id_alumno, codigo_carrera, pref, assigned) %>% dplyr::rename(codigo_carrera_assigned_int = codigo_carrera, pref_assigned_int = pref, assigned_int = assigned),
            ass_fin %>% dplyr::select(id_alumno, codigo_carrera, pref, assigned) %>% dplyr::rename(codigo_carrera_assigned_fin = codigo_carrera, pref_assigned_fin = pref, assigned_fin = assigned), by="id_alumno", all=TRUE)
```

```{r}
ass <- merge(ass, reg_int %>% dplyr::select(id_alumno, codigo_carrera, pref) %>% dplyr::rename(pref_assigned_fin_at_int = pref), by.x=c("id_alumno", "codigo_carrera_assigned_fin"), by.y=c("id_alumno", "codigo_carrera"), all.x=TRUE)
```

## Merge sources of data
```{r}
df <- merge(df, ass, by.x="MRUN", by.y="id_alumno", all.x=TRUE)
```

```{r}
df$assigned_oficial[is.na(df$assigned_oficial)] <- 0
df$assigned_int[is.na(df$assigned_int)] <- 0
df$assigned_fin[is.na(df$assigned_fin)] <- 0

df$codigo_carrera_assigned_int[is.na(df$codigo_carrera_assigned_int)] <- 0
df$codigo_carrera_assigned_fin[is.na(df$codigo_carrera_assigned_fin)] <- 0

df %<>% mutate(entered = ifelse(assigned_int == 0 & assigned_fin == 1 , 1, 0),
               entered = ifelse(assigned_int == 1 , NA, entered),
               left = ifelse(assigned_int == 1 & assigned_fin == 0, 1, 0),
               left = ifelse(assigned_int == 0, NA, left),
               changed_status_ass = ifelse(assigned_int != assigned_fin, 1, 0),
               changed_program_ass = ifelse(changed_status_ass == 0 & assigned_int == 1 &
                                             codigo_carrera_assigned_int != codigo_carrera_assigned_fin, 1, 0) )
```

```{r}
df$increased[is.na(df$increased)] <- 0
df$decreased[is.na(df$decreased)] <- 0

df$increased_valid[is.na(df$increased_valid)] <- 0
df$decreased_valid[is.na(df$decreased_valid)] <- 0
df$changed[is.na(df$changed)] <- 0
```

```{r}
saveRDS(df, paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/master_file_for_analysis_of_rct.csv", sep=''))
```

## Pre-processing of survey
```{r}
survey_mapping <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/survey/encuesta.csv", sep=''))
```

```{r}
survey_answers <- read_excel(paste(dropbox_dir,"Mistakes Structural/Data/Data_from_Mistakes_1/clean/survey-2022/survey answers/ans_surveymistakes.xlsx", sep=''))
```

```{r}
survey <- merge(survey_answers, survey_mapping, by.x="hash", by.y="llavespub")
```

```{r}
survey %<>% dplyr::select(mrun, QID8_11, QID131_11, QID12_2, QID95_1,
                          QID87_1, QID87_2, QID87_3, QID87_4, QID87_5, QID87_7, QID87_8, QID87_9, QID87_8_TEXT,
                          QID124, QID123_2, QID132_1,
                          QID25, QID26, QID125_1, QID125_2, QID125_3, QID125_4, QID125_5, QID125_6, QID125_7, QID125_7_TEXT,
                          QID161_1, QID162_1,
                          QID136_1, QID136_2, QID136_11, QID136_10,
                          QID137_1, QID137_2, QID137_11, QID137_10,
                          QID138_1, QID138_2, QID138_11, QID138_10,
                          QID25, QID26, QID26_4_TEXT,
                          QID27_1, QID27_2, QID27_5, QID27_6, QID27_5_TEXT,
                          QID150_2, QID25, QID27_1, QID27_2, QID27_5, QID27_5_TEXT, 
                          QID27_6, QID128_1,	QID128_2,	QID128_3,	QID128_4,	QID128_5,
                          QID134_1, QID134_2,	QID134_3,	QID134_4,	QID134_5,	QID135_1,	QID135_2,
                          QID135_3,	QID135_4,	QID135_5)
```

```{r}
survey %<>% dplyr::rename(prob_top_pref = QID8_11, prob_any_pref = QID131_11, codigo_carrera_top_true = QID12_2, puntpond_top_true = QID95_1,
                          bottom_true_any = QID124, codigo_carrera_bottom_true = QID123_2, puntpond_bottom_true = QID132_1,
                          puntpond_top_reported = QID161_1, puntpond_bottom_reported = QID162_1,
                          cutoff_prev_top_reported = QID136_1, cutoff_prev_bottom_reported = QID136_2, cutoff_prev_top_true = QID136_11, cutoff_prev_bottom_true = QID136_10,
                          cutoff_top_reported = QID137_1, cutoff_bottom_reported = QID137_2, cutoff_top_true = QID137_11, cutoff_bottom_true = QID137_10,
                          prob_top_reported = QID138_1, prob_bottom_reported = QID138_2, prob_top_true = QID138_11, prob_bottom_true = QID138_10,
                          knows_adm_mistake = QID25, reason_mistake = QID26,
                          single_choice = QID150_2)
```

```{r}
survey$codigo_carrera_top_true <- as.numeric(gsub("\\D", "", survey$codigo_carrera_top_true))
survey$codigo_carrera_bottom_true <- as.numeric(gsub("\\D", "", survey$codigo_carrera_bottom_true))
survey$codigo_carrera_single_choice <- as.numeric(gsub("\\D", "", survey$single_choice))
```

```{r}
survey %<>% mutate(bottom_true_any = ifelse(bottom_true_any == "Sí", 1, ifelse(bottom_true_any == "No", 0, bottom_true_any)))

survey %<>% mutate(reason_skip_top_true = ifelse(!is.na(QID87_1), QID87_1,
                                                 ifelse(!is.na(QID87_2), QID87_2,
                                                        ifelse(!is.na(QID87_3), QID87_3,
                                                               ifelse(!is.na(QID87_4), QID87_4,
                                                                      ifelse(!is.na(QID87_5), QID87_5,
                                                                             ifelse(!is.na(QID87_7), QID87_7,
                                                                                    ifelse(!is.na(QID87_9), QID87_9,
                                                                                           ifelse(!is.na(QID87_8), QID87_8_TEXT, NA ) ) ) ) ) ))))

survey %<>% mutate(reason_skip_bottom_true = ifelse(!is.na(QID125_1), QID125_1,
                                                    ifelse(!is.na(QID125_2), QID125_2,
                                                           ifelse(!is.na(QID125_3), QID125_3,
                                                                  ifelse(!is.na(QID125_4), QID125_4,
                                                                         ifelse(!is.na(QID125_5), QID125_5,
                                                                                ifelse(!is.na(QID125_6), QID125_6,
                                                                                       ifelse(!is.na(QID125_7), QID125_7_TEXT, NA )) ) ) ) ) ) )

survey %<>% mutate(reason_uninformed_requisites = ifelse(!is.na(QID27_1), QID27_1,
                                                      ifelse(!is.na(QID27_2), QID27_2,
                                                             ifelse(!is.na(QID27_6), QID27_6,
                                                                    ifelse(!is.na(QID27_5), QID27_5_TEXT, NA ) ) ) ))

survey %<>% mutate(reason_mistake = ifelse(is.na(QID26_4_TEXT), reason_mistake,  QID26_4_TEXT) )
#TODO: In the script of survey 2021 we have some cleanings that have to be applied here
# regarding the definition of top-true preferences
#TODO: Ask the RA to double check that we are identifying the correct questions
# from the survey in the script and which questions have skips or missing values

#Defining inconsistent types depending of the reason they give to skip top-true programs or bottom-true programs







#TODO: FINISH THIS AND ADAPT IT TO BOTTOM TRUE

# Translate labels of reasons to English (not an exact translation)
# dfs_2021 %<>% mutate(reason_strategic_1 = ifelse(reason_strategic_1 == "Postulé a mi carrera ideal en mi primera preferencia.",
#                                                  "(a) I did apply to my ideal program as a top-reported preference", ""),
#                      reason_strategic_2 = ifelse(reason_strategic_2 == "La probabilidad de que quede seleccionado en esa carrera es muy baja.",
#                                                  "(b) My admission probability to that program is too low", ""),
#                      reason_strategic_3 = ifelse(reason_strategic_3 == "La carrera me parece demasiado difícil y no creo que vaya a terminarla en caso de matricularme.",
#                                                  "(c) The program is too hard and I don't think I would be able to graduate from it", ""),
#                      reason_strategic_4 = ifelse(reason_strategic_4 == "No tengo los recursos económicos para pagar el costo de la carrera.",
#                                                  "(d) I do not have the monetary resources to pay for the program", ""),
#                      reason_strategic_5 = ifelse(reason_strategic_5 == "Para incluir mi carrera ideal tendría que excluir alguna de las carreras de mi postulación.",
#                                                  "(e) To include my ideal program, I would have to exclude some program from my list", ""),
#                      reason_strategic_6 = ifelse(reason_strategic_6 == "La decisión de dónde postular no dependió completamente de mí, y fue influenciada por otras personas (familia, amigos u otros).",
#                                                  "(f) The decision to where to apply did not depend only on me, and it was influenced by other people (family, friends, etc.)", ""),
#                      reason_strategic_7 = ifelse(reason_strategic_7 == "Pensé que al incluir esta carrera en mi lista de preferencias habría reducido mis posibilidades de selección en las demás carreras.",
#                                                  "(g) I thought that by including this program in my list I would have reduced my chances of being admitted to the other listed programs", ""),
#                      reason_strategic_8 = ifelse(reason_strategic_8 == "Dado que mi probabilidad de selección es muy baja, prefiero no postular y quedar seleccionado en una preferencia reportada más alta.",
#                                                  "(h) Given that my admission chances are too low, I prefer to do not list this program and being assigned to a higher reported preference", ""))
#
# # Create variable to identify inconsistent answers to the survey
# dfs_2021 %<>% mutate(inconsistent_type = 0)
# dfs_2021 %<>% mutate(inconsistent_type = ifelse((((reason_strategic_1 == "") | (reason_strategic_3 != "") | (reason_strategic_4 != "")) & (truth_teller == 1)), 1, inconsistent_type))
# dfs_2021 %<>% mutate(inconsistent_type = ifelse(((reason_strategic_1 != "") | (reason_strategic_3 != "") | (reason_strategic_4 != "")) & (misreport_order == 1), 1, inconsistent_type))
# dfs_2021 %<>% mutate(inconsistent_type = ifelse(((reason_strategic_1 != "") | (reason_strategic_3 != "") | (reason_strategic_4 != "")) & (misreport_exclude == 1), 1, inconsistent_type))
#


```

QID8 <- prob assignment top pref
QID131 <- prob assignment in any preference
QID12_2 <- top true
QID95_1 <- punt pond in top true
QID87_1 to _8 <- reason for not including top true in ROL
QID124 <- any other program outside ROL (bottom true) yes/no
QID123_2 <- bottom_true
QID132_1 <- score at bottom true
QID125_1 to _TEXT <- reason for not including
QID161_1 <- score in trop reported
QID162_1 <- score in last reported
QID136_1 <- cutoff last year top reported
QID136_2 <- cutoff last year last reported
QID136_3 <- cutoff last year top true
QID136_4 <- cutoff last year bottom true
QID137_1 <- cutoff top reported
QID137_2 <- cutoff last reported
QID137_3 <- cutoff top true
QID137_4 <- cutoff bottom true
QID138_1 <- prob top reported
QID138_2 <- prob last reported
QID138_3 <- prob top true
QID138_4 <- prob bottom true
QID25 <- knowledge about adm mistakes (yes/no)
QID26 <- reason
QID27_1 to_text <- if didn't know the requisite, why
QID150_2 <- single choice

```{r}
survey$responded_survey <- 1
```

```{r}
saveRDS(survey, paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/survey_processed.rds", sep=''))
write.csv(survey %>% dplyr::select(mrun, codigo_carrera_top_true, codigo_carrera_bottom_true), paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/top_and_bottom_true.csv", sep=''), row.names=FALSE) #had to get rid of sep=","
```

```{r}
survey <- readRDS(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/survey_processed.rds", sep=''))
tbp <- read.csv(paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/preprocessing_top_and_bottom_true.csv", sep=''))
#NOTE: the probabilities here might be outdated, so we will drop them and use the new clean data from Nacho
tbp %<>% select(!starts_with("prob"))

sur <- merge(survey, tbp %>% dplyr::select(-c(codigo_carrera_top_true, codigo_carrera_bottom_true)), by="mrun", all.x=TRUE)

sur %<>% dplyr::rename(puntpond_top_true_belief = puntpond_top_true.x, puntpond_top_true = puntpond_top_true.y,
                        puntpond_bottom_true_belief = puntpond_bottom_true.x, puntpond_bottom_true = puntpond_bottom_true.y)

sur %<>% mutate(puntpond_top_true = as.numeric(puntpond_top_true)/100,
                puntpond_bottom_true = as.numeric(puntpond_bottom_true)/100,
                puntpond_top_true_belief = as.numeric(puntpond_top_true_belief),
                puntpond_bottom_true_belief = as.numeric(puntpond_bottom_true_belief),
                diff_puntpond_top_true = puntpond_top_true - puntpond_top_true_belief,
                diff_puntpond_bottom_true = puntpond_bottom_true - puntpond_bottom_true_belief)
# Merge with the clean probs computed by Nacho

saveRDS(sur, paste(dropbox_dir,"Mistakes Structural/Code/Python/Cartillas/outputs/2022/survey_processed_with_probs.rds", sep=''))
```
